ARM: OMAP5: GPIO: do module reset coming from context loss
On OMAP5 ES 1.0, we have bug COBRA-1.0BUG00134 which
requires GPIO module to be reset when coming out of context loss
or for wakeup module, coming out of core domain OFF.

All the GPIOs are in L4PER domain except GPIO1 which is in WKUPAON
domain. Since, it is wkup domain it does not need it's context to be
saved/restored, So, GPIO1 gets special handling based off core PD
state.

Change-Id: I12a5c7f33015023300bf1b0fabbf236628456d6d
Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Pradeep Gurumath <pradeepgurumath@ti.com>
Signed-off-by: Huzefa Kankroliwala <huzefank@ti.com>