remoteproc: omap: fix api used by ipu to hold latency constraint
dev_pm_qos API are currently used to hold a latency constraint on
IPU. This effectively holds a constraint on the core power domain,
which includes IPU as well.

However, the pm core uses pm_qos APIs to control the latency of the
core domain and thus the C-states in OMAP. The two sets of interfaces
(pm_qos and dev_pm_qos) are de-coupled and thus cannot be used at the
same time to control latency values for core domain.

The remoteproc latency code has therefore been modified to use the
pm_qos API for holding latency constraints on ipu, so that the pm core
code has a consistent view of the constraints set on core power domain.
The dev_pm_qos API usage is retained for DSP, which resides in a
different power domain.

Change-Id: I17033a821279e520fff86deb6d7b16dd53a8b380
Signed-off-by: Subramaniam Chanderashekarapuram <subramaniam.ca@ti.com>
Signed-off-by: Paul Hunt <hunt@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>