input: keypad: smsc ece1099: ignore keypad scan interrupts
Interrupts can be generated by an edge detection in a GPIO
pin or a bus interface pins, also the interrupts can be
generated during the keyboard scan interface.

When the keyboard scan interface is enabled, each scan
column will trigger an irq event if the interrupt mask
is enabled, this is the reason why with the current
implementation a single key press was generating a
irq amount equal or at least the maximun number of
columns.

Now, before enabling the interrupt mask, enable the
keyboard scan interface and then clear any pending
interrupt.

Change-Id: I51c89cfcf8342c181f911e9d278b0c41eacbe9b4
Signed-off-by: Leed Aguilar <leed.aguilar@ti.com>