ASoC: Use sync version of runtime PM APIs
ASoC already uses sync version of runtime APIs for getting resources
but uses the async version to release them. Certainly, whether or not
the resources are released immediately may not matter in most cases,
but there are certain scenarios where this doesn't hold true.

External clocks that are not under control of clock framework can be
enabled at regular PCM operations (e.g. startup or shutdown) and the
order at which those clocks are enabled/disabled with respect to the
associated module enable/disable may matter. When using the async
version of runtime PM put, that order may not be honored as the
resource disable task is deferred.

One specific case is McPDM - twl6040 dependency. twl6040 provides an
external clock to McPDM. When McPDM module is getting disabled, it
must happen when the external clock is still present, otherwise the
module ends up in a bad state. This particular case needs the sync
flavor of runtime PM functions.

Change-Id: I53801a7fabb8fd31ba6a5b8a399c3aad81487517
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>