ARM: OMAP4+: DPLL: Add ability to dump DPLL registers.
Add ability to dump DPLLs registers before/after entering
suspend state.
The dumping level is controlled by module parameter
/sys/module/dpll44xx_54xx/parameters/dpll_dump_level
The solution is a flag based approach:
 0 - no dumps (default)
 1 - dump DPLLs registers before entering suspend
 2 - dump DPLLs registers before and after entering suspend
 4 - dump DPLLs registers around restore (before and after)
Combinations that can be used are:
1,2,3,4,5,6,7 - based on flags above depending on debug
conditions required.

NOTE: CONFIG_PM_ADVANCED_DEBUG is required to be enabled
to allow for the level of control.

Change-Id: I8311f79bcdd8b52f42c831cc98a31181081a61c5
Signed-off-by: Grygorii Strashko <grygorii.strashko@ti.com>