ARM: OMAP5: DPLL: Fix DPLL PER context definition
OMAP5 per DPLL offsets have a slight difference from
OMAP4 definitions. So, sync to spec for PER domain:
- Use OMAP5 registers offsets definition for dividers
 OMAP54XX_CM_DIV_H11_DPLL_PER_OFFSET
 OMAP54XX_CM_DIV_H12_DPLL_PER_OFFSET
 OMAP54XX_CM_DIV_H14_DPLL_PER_OFFSET
- Remove wrong access to CM_DIV_M6_DPLL_PER (0x4A008160) which is
not defined for OMAP5 (since tuple offset is 0, access will not
be performed for OMAP5).

We split the definition for OMAP4 and 5 and introduce inittime
logic to determine which DPLL set to choose from.

Change-Id: I4bbac3f6da42b24421e57f34dd533b602f5bfa36
Signed-off-by: Grygorii Strashko <grygorii.strashko@ti.com>