ARM: OMAP: wakeupgen: abort suspend if pending interrupts
There exists a time period in few 10s of milliseconds where
system would have interrupts disabled prior to attempting suspend
state. In this time interval, it is possible that we might
have an interrupt pending on the system which can prevent
MPU Power domain from hitting targeted state. In order
to handle this, we could check for pending interrupts
as part of the suspend logic, abort the suspend and handle
the event as needed and retry suspend after the pending event
is cleared. This ensures two things:
a) events which could have prevented targeted suspend state is cleared
b) such events are handled in a timely manner instead of attempting
suspend and waiting on another wakeup enabled event to wake and handle
all events.

Aborts suspend in case of pending interrupts with a report as follows:
Aborting Suspend: IRQ 69(gp timer)(OMAP-IRQ=37) Pending! Wakeup Enable: CPU0=yes, CPU1=no
Here, Linux IRQ number is 69, IRQ is registered with an action with
the name "gp timer", corresponding OMAP IRQ is 37(maps to TRM) and
the wakeupgen configuration is set to wakeup CPU0 and not CPU1.

In case the IRQ registered does not have a description registered,
The logic merely prints "unknown?" instead of the specific name.

However, it is a good practise to understand why such abort events
took place and handle the same in appropriate driver's suspend
handlers to allow a successful suspend at every attempt.

Minor change as part of this is to make the max_spi_reg to unsigned
int so that the type is the same as irq_banks for min() comparison
without build warnings.

Change-Id: I559d7802de079d842793b79f37f70234d8f16d86
Signed-off-by: Nishanth Menon <nm@ti.com>