ARM: OMAP: hwmod: Fix the module disable sequence
All IP blocks/modules which support "Auto" and "Disable" modulemode
configuration, need the module to be in FUNCT (functional) state before
it can be DISABLED.
An "Auto" (0x1) to "Disable"(0x0) modulemode change will not transition
them from IDLE to DISABLED state. Instead they should be first transitioned
from IDLE to FUNCT and then from FUNCT to DISABLED state.
Refer to Figure 9 "Idle Protocol State Chart" of the OMAP4 PRCM
functional spec document to know more.

The way to transition a module from IDLE to FUNCT state is to force
the clockdomain, which the module belongs to, to wakeup (SW_WKUP).

Follow this sequence in the _idle and _shutdown functions in the
hwmod framework.

Change-Id: Iec43fb93ee7fb00a51c942e8a1471f7635839c46
Reported-by: Cedric Vamour <c-vamour@ti.com>
Signed-off-by: Rajendra Nayak <rnayak@ti.com>