OMAP5: sEVM board file: Fix USB host & ethernet mux settings
Due to GPIO bug OMAP5430-1.0BUG01667 in OMAP5 ES1.0, a glitch is
sometimes observed on the USB & ethernet reset GPIO pins.
WA is to change direction to input. For this WA to be effective
the mux configuration needs to enable INPUT_PULLUP. This will
have no effect when configured in o/p mode but is required when
pin is changed to i/p during OSWR/OFF mode.
This along with "OMAP5: GPIO: WA for glitch on USB host/Ethernet
reset GPIO pins" can be reverted when this is resolved in ES 2.0.

Also omap_mux_init_gpio is a better API to use for GPIO pins.

Change-Id: Ida4c8bbbec37ad29b2ce10430991ae0430e7257c
Reported-by: Mantesh Sarashetti <mantesh@ti.com>
Signed-off-by: Huzefa Kankroliwala <huzefank@ti.com>