TI thermal framework: Dummy cooling device to update userspace
This patch introduces a dummy cooling device to update
the userspace about the cooling level. A module parameter
user_cooling_level will provide the cooling level information.
Whenever there is a change in the cooling level, this parameter
is updated and change notification is sent. It is upto the
userspace policy to act on the cooling level and take appropriate
cooling actions.

The user_cooling_level can found at
/sys/devices/platform/user_cooling_device/level

Change-Id: Idff140c37935fa7bc4e381309a3e1d5b084c5ffc
Signed-off-by: Radhesh Fadnis <radhesh.fadnis@ti.com>