serial: omap: uart: fine grained control of the rts line
This patch is based off this patch.

commit 9836877d7d20992ca0e65259bc5c9f8b0ff4b193

Implement correctly the Auto-RTS uart feature.
RTS line should be kept pulled high, when uart is not in use.

Provide infrastructure to pass the RTS line pad mux from board file.

Also RTS control has to be very fine granined as immediately after
RTS goes low, other end can start sending the data.

So make RTS low only when we know uart is ready to handle RHR/Rx interrupts.
Run-time-pm callbacks are the right place to have this RTS control implemented.

System-wide-suspend is a special case that needs to be handled, as the whole
setup gets repeated for tty layer where request_irq() happens again.
So set_termios is the right place if we are coming out of a system-wide suspend,
to control RTS line for flow control.
Author: Vikram Pandita <vikram.pandita@ti.com>

Change-Id: Ife64e3fa3ca8c6e421e44d348c34f8e60d2a7dfe
[Ported from K3.0 commit 9836877d7d20992ca0e65259bc5c9f8b0ff4b193]
Signed-off-by: Ruchika Kharwar <ruchika@ti.com>
Signed-off-by: Vikram Pandita <vikram.pandita@ti.com>