OMAP: Serial: Enable rts_mux_driver_control for uart5
uart5 is connected to the bluetooth IP. This setting enables
the rts_mux_driver_control which allows the rts line to pulled up
appropriately to force the bluetooth chipset to assert the CTS
before sending data. This is a workaround to avoid a character loss
on the UART RX pins when it is idle.

Change-Id: I0b35a31cefbf2b00454b0fe70dd4c859c656ba27
Signed-off-by: Ruchika Kharwar <ruchika@ti.com>