ARM: OMAP5: Update workaround for sync-timer and watchdog errata
Errata Titles:
- GP and WD Timer Returns Incorrect Time Due to Dynamic Clock Gating (i767)
- 32K Timer Returns Incorrect Time Due to Dynamic Clock Gating (i768)

A static dependency between the MPU and WKUPAON power domains was added to
workaround the above errata for the synctimer and WD timers. This workaround
has been updated to only apply the static dependency for OMAP543x ES1.0. For
OMAP543x ES2.0 the sync-timer errata (i768) will be fixed and given that the
Linux watchdog driver does not read the watchdog counter, no workaround is
currently needed for the watchdog.

Signed-off-by: Jon Hunter <jon-hunter@ti.com>