ARM: OMAP2+: pm_debug: report current power domain state counter
If the expected power domain state is not the same as current
powerdomain state, then we currently print the count of the
expected power domain state(in addition to reporting
the mismatch).
However, in this mismatched state, count data provided is
spurious as the user is interested in current state data.

So, provide user always with current state count information

Change-Id: Ieebc146eab36c8a1b94d7143509d6104bbdccf97
Signed-off-by: Mahesh Renduchintala <mahesh@ti.com>