GPIO: OMAP: Fix level+edge interrupt handling
In the current GPIO driver we clear the edge interrupts
right away whereas the level interrupts are handled after
the GPIO client handler is run. For GPIOs programmed for
both level+edge interrupts the IRQ handler does same as in
level case.

This worked fine as long as MPU->L4PER static dependency was
enabled in which case GPIO blocks which are in L4PER would
transition along with MPU and in a way never need swakeup
events to wakeup. Also swakeup event would not happen for
level only GPIOs. With no dependency we see that for GPIO
like one used by touchscreen programmed for level+edge the
GPIO block idles frequently and if there is a swakeup event,
in IRQ handler since we do not immediately clear the IRQSTATUS
(for level+edge we follow only level sequence where it gets
cleared after handler runs) the swakeup never gets deasserted
and the block never idles again. As a result of this when module
is explicitly disabled it gets stuck in transition.

The key to de-assert sWakeup is to ensure that the status is
cleared before the interrupt is disabled else it will cause
sWakeup to stay high and the module to get stuck in transition.

This patch ensures that for edge & level+edge condition the
status is cleared right away in interrupt handler. The level
condition is handled as before.

Change-Id: Ieda1715a96cdf8bc5563ad4392f5718426f26349
Signed-off-by: Huzefa Kankroliwala <huzefank@ti.com>