GPIO:OMAP5:Fix incorrect restore of GPIO pins in WA
The WA for GPIO pins was moved to runtime suspend/resume
handler from system suspend/handler.
However the direction of pins was not restored correctly
since the place where it was put in runtime resume handler
would not get executed always.

Fix this to ensure that the WA is applied properly

Change-Id: Iea1dcbb05eeb8ad11ebf8ea623d4d847fa7ded56
Signed-off-by: Huzefa Kankroliwala <huzefank@ti.com>