OMAP: Serial: Populate noirq handlers for DEV_PM
This patch removes the enable_irq() and the disable_irq() calls from
the resume and suspend handlers respectively.
Instead, the resume_noirq and suspend_noirq handler handlers are initialized
in the DEV_PM structure. Since the other *_noirq handlers are not used, they are
not employed.
The above change necessitates the request_irq and free_irq to move to the
probe() and remove() functions respectively. "enable_irq" is called during resume
from the pm framework hence, the irq cannot be freed in the "shutdown" function.

Change-Id: I4c7cd95553dfc40477ba0421f2989154cff3d3b7
Signed-off-by: Ruchika Kharwar <ruchika@ti.com>