remoteproc: omap: control mailbox states as part of rproc pm
The omap remoteproc driver, being a user of mailbox, needs to take
care of enabling/disabling mailbox properly in the pm resume/suspend
paths. This will allow the mailbox driver to save its context and
release any constraints in the disable path, and restore them back
in the enable path.

Change-Id: I0e33bf57b17eacc23e7374f07351b8e92626f3b2
Signed-off-by: Subramaniam Chanderashekarapuram <subramaniam.ca@ti.com>