OMAP5: GPIO: WA for glitch on USB host/Ethernet reset GPIO pins
In OMAP5 GPIO pins 172 & 173 are reset pins for ethernet and
usbhost module respectively. Due to h/w bug OMAP5430-1.0BUG01667
in ES 1.0 sometimes there is a glitch observed while going to
OFF mode resulting in reset of these modules. This patch is a
WA for this issue where the pin directions are changed to input.
Along with this the MUX also needs be modified.
This needs to be reverted since the issue is not observed on
ES 2.0

Change-Id: Id05d8204a30813dcd8dc0b45c32f3cd38205c328
Reported-by: Mantesh Sarashetti <mantesh@ti.com>
Signed-off-by: Huzefa Kankroliwala <huzefank@ti.com>