OMAPDSS: DISPLAY: Optimize the omap5sevm panel timings
Description: This patch updates the panel timings and the
pixel clock from the DSI confifuration tool with new set of
DISPC & DSI porch values. The PLL configuration is part
of the board file change and can be found in the this patch
http://review.omapzoom.org/#/c/25894/.

This patch along with the board file patch should be used
together.

Change-Id: I2d02c9de97424e01864d218235490bec4d2c2048
Signed-off-by: Sreenidhi Koti <sreenidhi@ti.com>
Signed-off-by: Archit Taneja <archit@ti.com>