I2C: OMAP: replace runtime_put with the sync equivalent
pm_runtime_put is a lazy operation which disables clocks from PM
runtime workqueue. However after a I2C transfer is complete, it
is required that we do not have I2C clock active to allow for
lower power state transition. If the clocks were active, this
would prevent opportunistic transitions (like that of CPUIdle)
to take place. Hence, we replace this with the pm_runtime_put_sync
which guarentees that the clock is disabled at the end of the
transfer.

Change-Id: I4e131ae642e44e84e46531a43422ea7aa18e35a7
Signed-off-by: Grygorii Strashko <grygorii.strashko@ti.com>