OMAPDSS: HDMI: Fix Race condition in HPD worker thread
We continously see dss errors stating window width and video pipe width are not correct.
Simple way to reproduce issue is to have hdmi as default and start live wall paper and
plug on HDMI cable.

It is seen that we have lot of compositions in DSSCOMP queue. But these not redirected to the
display since the device state is not active when checked by DSSCOMP. Since HDMI
is default display all compositions are always directed to HDMI. This is the reason
why LCD + HDMI scenario we cannot reproduce this issue.

When HPD interrupt is received we come into worker thread. Here first thing we do is
enable HDMI. This sets the device state to active. Then we try to read edid. The
compositions which were avoided to be routed suddenly start routing. At this moment
edid is not read so device wxh is 0x0. So compositions created are going to fail
for panel wxh check. The log is so verbose that the edid read work which was
queued is fully preempted and we get a kernel crash due to GFX dump.

In theory we dont need hdmi to be on for edid to be read. So it is always a good
idea to read edid first. get proper device wxh determined and then turn on the
display.

Current solution logic:
1. Look if max edid read tries are failed. If so stop the state machine
2. If we are starting the state machine and retries are within limit read edid
3. if edid read is ok then change the state to edid read - 1 so that when we enqueue the work
again we get correct edid read state and notify userspace.
4. if edid is read and hdmi is off then turn on the hdmi and cec
5. Do not turn on HDMI when it is in suspended state

Change-Id: If167d0847dfa3821898fd34ccdc2480960b15a7d
Signed-off-by: Dandawate Saket <dsaket@ti.com>