ARM: OMAP5: VC: increase retention voltage for ES1.0
ES1.0 samples are not completely tested for Retention voltage
of 700mV, hence increase the retention voltage to 800mV to
work with untested and tested ES1.0 samples.

OMAP543x Data manual operating condition addendum revision 0.3
does not contain this information yet and is expected to be
updated in follow on revision.

As per the expectation, ES2.0 is supposed to have proper values
and hence split out the current ES1.0 values in preperation for
ES2.0 values as and when they are available. Meanwhile, for
OMAP revision > ES1.0, we will report a WARN.

Change-Id: I13348960376b38ffb05c12f89d525017db9f6631
Signed-off-by: Nishanth Menon <nm@ti.com>