OMAPDSS: HDMI : Fix Race condition between panel disable and DSSCOMP queue
Issue:
When setmode, disable, suspend is called on hdmi panel we see a crash
when interupt routine is called in case HDMI is default. This issue is
not seen on LCD+HDMI combination since LCD will hold the DSS clocks.
But in HDMI case since it is the only device on clock put clocks
are immediately released. So if ISR is triggred oe we are in ISR
we have random crashes. Releasing the DSSclock and disabling hdmi
was a forceful event without considering activity on HDMI.

Solution:
1. In disable panel call we communicate first that HDMI state is disabled
2. We wait for some time so that compositions are flushed out.
3. It time is empirical but idea is to make sure DSSCOMP queue function takes
note of the DSSDEV disabled state and stop queuing more compositions.
4. Then black the manager so that all the compositions are flushed from gralloc queue.
If you dont do this then GFX crash will happen for the composition which is currently
active.
5. Now disable the panel and release the clocks

Change-Id: Iaafbec63caf981e95dfe188e85c72286b4e0e110
Signed-off-by: Dandawate Saket <dsaket@ti.com>