HACK: OMAP5: cpuidle: Disable CPU idle during system suspend/hibernation
It was found that if CPUIdle is active during system suspend/hibernation,
it may cause overall system lockup. This occurs despite the fact
that disable_hlt/enable_hlt is called in OMAP implementation of
platform_suspend_ops begin and end callbacks. Rootcause is unknown
at this point in time.

However, by disabling CPUIdle as early as possible in the system
suspend/hibernation sequence, the system lockup is no longer
seen. We introduce this hack in the same approach as followed by
cpu_hotplug module.

Change-Id: I1ab8c7345de859854f2e33edc7e2ac19c2c8449e
Signed-off-by: Grygorii Strashko <grygorii.strashko@ti.com>