OMAP4+: omap mux init: enable pull down for fref_clk_ioreq pad
Sysboot bits define if fref_clk_ioreq pad is input or output signal.
fref_clk_ioreq is used as clock request for auxclk0, and if request is
asserted (which could be due to not enabling pull down) it can keep
auxclk0 active, if this clock is not explicitly disabled. This also
leads to PWR_REQ signal is not de-asserted in OFF mode.

Benefit of doing this using sysboot setting is that platform
configuration changes are mandatory to be reflected on sysboot. This
allows us to not depend on board files registering information, instead
depend on actual h/w detected configuration for mux direction and making
board files independent of the change if crystal is used (TI recommended
configuration)

If fref_clk_ioreq is used as input, then enable pull down to let OMAP
idle unused auxlock0.

Change-Id: I9dd338c8e8e65e1f1e185d297981491c8bbf620c
Signed-off-by: Oleksandr Kozaruk <oleksandr.kozaruk@ti.com>