ARM: OMAP2+: Make omap_usec_to_32k API common
The omap_usec_to_32k API is intended to be used for conversation
from microsecond value to number of 32kHz clock cycles. And now,
this, actually, common function is defined as private function in VC
module.

Hence make it common, so other modules will be able to use it too.

Change-Id: Id6ba932442639fa5765e8ee3500153c5f2cca70b
Signed-off-by: Grygorii Strashko <grygorii.strashko@ti.com>