ARM: OMAP5: hwmod data: bb2d module related data only for ES2.0
bb2d (bit blit 2d accelerator) is not expected to be functional
in the OMAP5 ES1.0 silicon. Hence, even the ES1.0 TRM do not have
any mention of this module. ES2.0 however is expected to have a fully
tested and functional and probably newer IP block intergrated.
So, skip bb2d related data while we are dealing with ES1.0 and add
it back along with and when ES2.0 is detected. We might update
the ES2.0 data once we have finalized information. For the timebeing
carryover the existing information.

This patch is done my manually editing the existing omap5 hwmod data
file and not by fixing the autogen scripts because of the mismatch
between the data files in the kernel and the autogenerated output.

Change-Id: Ia01fa49ca470f3c928ebe809f494d44d1caa5bbd
[nm@ti.com: runtime detection]
Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Rajendra Nayak <rnayak@ti.com>