OMAPDSS : HDMI : Use correct mode and code when HPD happens
Currently we always fall back to 640p when driver->enable
function is called. If 1080p TV is connected omapfb selects
panel as 1080p and allocate 1080p Framebuffer.
So if we pull out the cable during intial andriod bootup logo
there is no listner on the userspace to select best mode
and panel is turned on with 640p. So the 1080 content now start
routing to 640p causing DSS to error for wrong panel wxh.

Solution:
Try to first fall back to previous selected mode+code.
If we are not able to do so we will select 640p.

Change-Id: Ie5265764d9603d80b3fcfecc33715c1b8f33841e
Signed-off-by: Dandawate Saket <dsaket@ti.com>