GPIO: OMAP: Always scan gpios during runtime resume
This allows the driver to generate interrupts for GPIOs that can't
wakeup but have changed state during runtime suspend. We cannot
depend on the decision based on no need to restore, as the system
state might change and pending events could gather up.

This also means that during suspend, we cannot assume that we should
save context ONLY if we are entering device wide OFF mode, as
the powerdomain can independently transition

Since bank's power_mode is no longer used, remove it.

Change-Id: I917b6e15ba7ddac385b003088967855d1fb5d9fa
Signed-off-by: Tero Kristo <t-kristo@ti.com>