ARM: OMAP4+: mux: reconfigure io chain after detected GPIO event
Recommended way to clear IO chain wakeup events on OMAP5 is that for
each detected wakeup events, the WAKEEN bit should be disabled and an
io chain reconfigure is executed. For GPIOs this effectively means
that the WAKEEN bit must be toggled, and the reconfigure must be
executed twice so that the wakeup remains active.

Change-Id: Iea945871c816a1aafb31f68b7f4d30cc25b8a516
Signed-off-by: Tero Kristo <t-kristo@ti.com>