rpmsg: omx: lookup omxserv by device name instead of minor number
The rpmsg_omx driver is reprobed upon a remoteproc recovery, but the
cdevs exposed to the userspace are reused. The minor number is being
used currently as the identifier to look up the original created
omxserv. However, this is error prone since it assumes that the order
of the omxserv creation is always the same.  This issue can be avoided
and handled better by matching the name of the omxserv with the rpdev
name.

Also, it is not safe to create a new minor device number using idr
for every recovery, as it might give a different value than the one
generated during character device creation. This may cause issues when
destroying the character device. Hence, the minor device number/idr is
not destroyed during recovery and is reused.

Change-Id: I0d553e09dc610bd8d4d4baff58cfe641ee3f02f5
Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>