TI-TFW: omap-bandgap: Add support for Core Cpuidle
Omap-bandgap temperature sensor needs to be on as long
as the device is on. But this will prevent the core from
idling. Implemented api's to request and release omap-bandgap
clocks. These api's have to be called from the PM idle notifiers
to Idle the bandgap module during Core Low power transitions.

Change-Id: I37c82757c9ab22198e2d2fa33e60080556bb0057
Signed-off-by: Radhesh Fadnis <radhesh.fadnis@ti.com>