OMAP5: Battery: Detect battery to enable reporting
When switch S5 is turned off, battery positive is
disconnected, however still fuel guage TS pin is connected
to battery thermistor, resulting in voltage reported
as as 3.7V which is the directly coming from LDO.

Hence to detect if battery is present or not,
first check for battery current (through current sense
resistor). If reported current is zero, battery
is full or not present. In this case get battery
voltage (through voltage sense resistor) if the voltage
is 3.7-3.9V it is coming from the LDO and greater
than 3.9V tis from battery.

Change-Id: I5a946836123dbdac43c318502478677defed75eb
Signed-off-by: Madhusudan C M <c-madhu@ti.com>