misc: emif: use restart if power_off not present when out of spec
Some machine or kernel variants might have missed implementation
of power off handlers. We DONOT want to let the system be in
"out of spec" state in this condition. So, WARN and attempt
a machine restart in the hopes of clearing the out-of-spec
temperature condition.

NOTE: This is not the safest option, but safer than leaving the
system in unstable conditions.

Change-Id: If4a5c741169432d12b89d66124ebe1dd2e6435e9
Signed-off-by: Nishanth Menon <nm@ti.com>