ARM: OMAP5: Cpuidle: Program MPU PD to ON state after cpuidle
MPU Powerdomain needs to be programmed to ON state after coming
out of cpuidle. Otherwise we can end up in invalid c state combination.
Also idle notifier logic would fail as it would use mpu powerdomain state
for calling idle notifiers.

Also update/read of mpu_state_vote and setting of mpu powerdomain
is not done as a atomic operation and hence can race. So protect this
with a spin_lock.

Change-Id: I0336f4849d73916e8d5968ac7fd52deb5e9325a4
[nm@ti.com: gerrit override, potential minor updates]
Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Vishwanath BS <vishwanath.bs@ti.com>