OMAP4+: HSI: WA: Handle the IO chain wakeup in ISR
There is no way to check the WAKEUPEVENT of the CAWAKE PAD which
is the reason of wakeup and of the ISR call.

So if no HSI IRQ has been capture, it should be an ISR called for
the IO chain wakeup and so we handle it.

TODO: Use a future API allowing us to check the WAKEUPEVENT for
the CA_WAKE PAD.

Change-Id: I140f2bde7b7ffbec0bacf850bf0802af45e7c71f
Signed-off-by: Cedric Poignet <c-poignet@ti.com>