ARM: OMAP5: CPUidle: do not overwrite cpu_idle_driver data
Currently we do the following:
a) omap5_idle_driver is registered twice (once per device)
b) The same driver data structure is modified from scratch after:
	b.1: registering the driver
	b.2: CPU0 device is registered  as well
At b.2 CPU0 is allowed to be controlled by cpuidle governors.
But then, we start changing the state_count etc get modified
on omap5_idle_driver structure for CPU1.

This is not good on a live data structure. So use a dummy
structure for the life of the operation and discard it
once done. update the drv structure and register at the
point of the first CPU only.

Change-Id: I188fb1779f12559b4d4e6eee7ecc5d485e50de22
Signed-off-by: Nishanth Menon <nm@ti.com>