usb: musb: omap2430: don't loop indefinitely in interrupt.
When called during resume_irqs, omap2430_musb_set_vbus() is run with
interrupts disabled,  In that case 'jiffies' never changes so the loop
can loop forever.

So impose a maximum loop count and add an 'mdelay' to ensure we wait
a reasonable amount of time for bit to be cleared.

This fixes a hang on resume.

[Ruslan]: cherry-picked from upstream
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux.git;
a=commit;h=12a19b5f836c1a1b838e668f9aa804d017d8c58b

Change-Id: I02aab4597014361a9528d3f90a8f3144057ef51a
Signed-of-by: NeilBrown <neilb@suse.de>
Signed-off-by: Felipe Balbi <balbi@ti.com>
Signed-off-by: Ruslan Bilovol <ruslan.bilovol@ti.com>