HACK: OMAP5: Program L4PER in CSWR instead of OSWR to avoid glitches for ES1.0
To avoid the glitches seen on the GPIO pins when L4PER transition to
OSWR, program the L4PER to CSWR instead of OSWR while in the idle path.

We suspect the issue to be related to the (Bug ID: OMAP5430-1.0BUG01667).
A Defect is opened to track this issue OMAPS00276448

Change-Id: Id8664299b7fadccfae630e3069f1701e96f0ef53
Signed-off-by: Nicole Chalhoub <n-chalhoub@ti.com>