drivers: nfc: Send proper ret val if reg to ST.
Adding proper return if already registered with ST
in this case send EAGAIN.
In cases where COM module was not there, we see
kernel panic when we do not return proper value
from open function. Since NFC driver was already registered
when we try to register again there was need to send
proper return value.

Change-Id: I979b80af675547254a8c28ead30ab1c41889955a
Signed-off-by: bvijay <bvijay@ti.com>