omap: mmc: Remove duplicate map/unmap of dma segments
With introduction of .pre_req callback, the dma segments
are already mapped by the time the .request callback
is called. So, there is no reason to map the dma segments
again.

Similarly, the unmapping of dma segments is taken care of
by the .post_req callback. So, no need to unmap dma segments
in the interrupt service routine of mmc.

Change-Id: If29a5d873ac2f26d11fe1f64eb6e1639229feacc
Signed-off-by: Viswanath Puttagunta <vishp@ti.com>