HACK: ARM: OMAP4+: board-serial: disable autosuspend for uart5
uart5 is used by BT/FM and autosuspend for this uart should be set by
shared transport driver.

without this hack , suspend/resume of platform is prevented before
BT is initialized.and this causing issues with OFF mode.

there is problem with uart5 during the system wide suspend handler.
system crashes because the get_sync does not appear to be enabling
the interface clocks from few minutes of debug.

Temp hack should be removed once investigation completes and correct
fix should be taken.

Change-Id: Ib2ad25965d83efbc896718cceefa9277f0f7fd04
Signed-off-by: Vishal Mahaveer <vishalm@ti.com>
Signed-off-by: Girish S G <girishsg@ti.com>