omap-hsmmc: fix errata i761: Card Error Interrupt May Not Be Set Sometimes
Inherited from commit 801bfd921ee0c55614f6aa6892ee3c0381705a15 in p-android-omap-3.0

Errata description:
Due to a bad behavior of an internal signal, the Card Error interrupt bit
MMCHS_STAT[28] CERR may not be set sometimes when an error occurred in the
card response.

Workaround:
After responses of type R1/R1b for all cards and responses of type R5/R5b/R6
for SD and SDIO cards, software must read two registers: MMCHS_RSP10 and
MMCHS_CSRE. When a MMCHS_CSRE[i] bit is set to 1, if the corresponding bit at
the same position in the response MMCHS_RSP10[i] is set to 1, the host
controller indicates a card error and software should proceed in the same way
as if a CERR interrupt would have been detected in the MMCHS_STAT register.

Note:
This errata is applicable for omap44xx.

Signed-off-by: Semen Protsenko <semen.protsenko@ti.com>

Resolved merge conflicts and adapted to p-linux-omap-3.4

Change-Id: Ifb283be671ed8d474a65e485dc2e9aa33ab8d3e8
Signed-off-by: Ranganath Krishnan <ranganath@ti.com>