ARM: OMAP3+: CPUIdle: build file only if defined
We do not need to attempt a build of cpuidle files if CONFIG_CPU_IDLE
is NOT defined. So, move the empty init functions into static inlines
and ensure that build of cpuidle files take place only if defined
in .config.

Also fixes:
arch/arm/mach-omap2/cpuidle5xxx.c:28: warning: 'mpu_lock' defined but not used
when CONFIG_CPU_IDLE is not defined

Change-Id: I95df80353cbdde908a630501a0cbb4272138cde1
Reported-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>