GPIO: WIFI: configure pull-down for IRQ line
The GPIO WIFI IRQ line currently configures the mux for this
pad as OMAP_PIN_INPUT. This is causing spurious interrupts in
OFF mode when the firmware for chip is not loaded. Since the
IRQ line is rising (from low to high) the pad should be
configured for pull down (OMAP_PIN_INPUT_PULLDOWN).
With this change we dont see any spurious wakeup events from
OFF mode.

Change-Id: I212fc45e29eb22e3e97925dffe84311db7394dbf
Signed-off-by: Huzefa Kankroliwala <huzefank@ti.com>
Signed-off-by: Pradeep Gurumath <pradeepgurumath@ti.com>