HACK: OMAP5: GPIO: Do SWSUP idling of GPIO2-8 modules
GPIOs can get stuck in transition during  the idle path while disabling
them, the following scenario is happening:
- Disable MPU Interrupts in the idle path
- a swakeup event happens on one GPIO
- the interrupts are disabled so the GPIO swakeup will not be served
- call the gpio notifiers to disable the GPIO module
- the GPIO is going to be stuck in transition

To avoid the GPIOs stucking in transition, we manually bring GPIO2-8
in and out of idle.

REVISIT: This is a temporary WA until the correct sequence is identified.
The issue is tracked with the following defect OMAPS00276446

Change-Id: Ic1e9d44752193955e29955b9045e66139dd53619
Signed-off-by: Nicole Chalhoub <n-chalhoub@ti.com>
Signed-off-by: Huzefa Kankroliwala <huzefank@ti.com>
Signed-off-by: Girish S G <girishsg@ti.com>