ARM: OMAP5: reboot support: Hack around i744 erratum
Erratum i744 for OMAP5 exposes a bug where device failure
is seen after a warm reset due to the fact that HSDIVIDER
ratio is messed up for a few cycles.

Recommended WA is to connect sys_nreswarm to reset_in pin
to force a warm reset to become a cold reset.

We instead trigger a cold reset in the case of ES1.0 samples
to work around this limitation in hardware. Expectation
is for future ES revisions to have appropriate fix. The WA
is printed to log hoping to have it in persistent trace
as needed.

Change-Id: I1d41eb069be3b19fa012344e2f1666db2b51d31c
Signed-off-by: Nishanth Menon <nm@ti.com>