regulator: TWL6030: add clk32kaudio as regulator
Add the 32KHz audio clock as a regulator as it
it may be used as a clk source for some external
chips. By adding it, we can enable and disable it
using the regulator framework.

This patch was ported from K3.0
commit: 4e93124a83ba1bd4c1cce2f9e8102ec7ace3c684

Additionally was added entry to the twl_of_match[] table.

Change-Id: Icabbdc6ffdf968f3e5a6ac16db1d14f250dab61f
Signed-off-by: Axel Haslam <axelhaslam@ti.com>
Signed-off-by: Oleksandr Dmytryshyn <oleksandr.dmytryshyn@ti.com>