ARM: OMAP3+: clockdomain: Enable usecounting for clocks
When clockdomain is controlled from clock framework, there is no
module mode information available for clocks and hence usecounting
is skipped. When one of the clocks in a clock domain is controlled
via clock framework and rest via hwmod, we might end up in a situation
where clock domain is disabled with some modules active which can lead
to aborts. This issue can be solved by including clockframework as well
in clockdomain usecounting.

Now when a optional clock is enabled via clock framework, it will be
usecounted in clockdomain framework. So when optional clock is disabled,
the clockdomain will be disabled only when rest of the modules in that clock
domain have idled.

Suppose a clock domain has some SW controlled modules and some HW controlled
modules. In this scenario, usecounting is done only for SW controlled modules.
However when usecount becomes 0 for that clock domain, the clockdomain will not
be disabled. Rather it will be kept in HW_AUTO mode (as it has HW controlled
modules). Hence there should not be any issue with HW controlled module's
functionality.

Change-Id: I73ef17817d4cd639f45b6aef6d642b06027e0432
Signed-off-by: Vishwanath BS <vishwanath.bs@ti.com>