OMAP5 sEVM: MODEM: WA: Fix WU event clear using another GPIO API
Only omap_mux_init_gpio() registers the GPIO PAD as WAKEUPENABLE
to allow the framework to clear the WAKEUPEVENT when necessary.

So we use it only when necessary after modem detection.
If we do it before, there is no way to unregister the GPIO from
the WAKEUPENABLE list.

TO DO: Use an API to unregister the PAD from wakeupenable list
maintained in mux.c.

Change-Id: Ifb54a3dad55d93e4a17c9998886da3e34a4d7d02
Signed-off-by: Cedric Poignet <c-poignet@ti.com>